requirement.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aniso8601==7.0.0
APScheduler==3.6.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
channels==2.4.0
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==2.5.0
Django==3.2
django-appconf==1.0.4
django-channels-graphql-ws==0.8.0
django-fsm==2.7.1
django-fsm-log==2.0.1
django-graphiql==0.4.4
django-subscriptions==2.1.1
graphene==2.1.8
graphene-django==2.15.0
graphene-subscriptions==1.0.2
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-ws==0.3.1
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
msgpack-python==0.5.6
multidict==5.1.0
ndg-httpsclient==0.5.1
promise==2.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
python-telegram-bot==13.4.1
pytz==2021.1
redis==2.10.6
requests==2.25.1
Rx==1.6.1
service-identity==18.1.0
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
yarl==1.6.3
zope.interface==5.3.0
